added frost green color preset and implemented mod menu mod config compatibility

+++ src/client/java/net/spellcraftgaming/rpghud/mixin/RenderOverlayMixin.java
@@ -41,7 +41,7 @@ public class RenderOverlayMixin {
     @Inject(at = @At("HEAD"), method = "tick")
     private void tick(CallbackInfo info) {
         ++this.ticks;
-        ;
+
     }
 
     @Inject(at = @At("HEAD"), method = "renderStatusBars", cancellable = true)
@@ -60,7 +60,7 @@ public class RenderOverlayMixin {
         if (playerEntity == null) {
             return;
         }
-        int i = MathHelper.ceil((float)playerEntity.getHealth());
+        int i = MathHelper.ceil(playerEntity.getHealth());
         boolean bl = this.heartJumpEndTick > (long)this.ticks && (this.heartJumpEndTick - (long)this.ticks) / 3L % 2L == 1L;
         long l = Util.getMeasuringTimeMs();
         if (i < this.lastHealthValue && playerEntity.timeUntilRegen > 0) {
@@ -77,22 +77,22 @@ public class RenderOverlayMixin {
         }
         this.lastHealthValue = i;
         int j = this.renderHealthValue;
-        random.setSeed((long)(this.ticks * 312871));
+        random.setSeed(this.ticks * 312871L);
         HungerManager hungerManager = playerEntity.getHungerManager();
         int k = hungerManager.getFoodLevel();
         int m = scaledWidth / 2 - 91;
         int n = scaledWidth / 2 + 91;
         int o = scaledHeight - 39;
         float f = Math.max((float)playerEntity.getAttributeValue(EntityAttributes.GENERIC_MAX_HEALTH), (float)Math.max(j, i));
-        int p = MathHelper.ceil((float)playerEntity.getAbsorptionAmount());
-        int q = MathHelper.ceil((float)((f + (float)p) / 2.0f / 10.0f));
+        int p = MathHelper.ceil(playerEntity.getAbsorptionAmount());
+        int q = MathHelper.ceil((f + (float)p) / 2.0f / 10.0f);
         int r = Math.max(10 - (q - 2), 3);
         int s = o - (q - 1) * r - 10;
         int t = o - 10;
         int u = playerEntity.getArmor();
         int v = -1;
         if (playerEntity.hasStatusEffect(StatusEffects.REGENERATION)) {
-            v = this.ticks % MathHelper.ceil((float)(f + 5.0f));
+            v = this.ticks % MathHelper.ceil(f + 5.0f);
         }
         if(RenderOverlay.shouldRenderVanilla(HudElementType.ARMOR)) {
 	        client.getProfiler().push("armor");
@@ -147,8 +147,8 @@ public class RenderOverlayMixin {
 	        if (playerEntity.isSubmergedIn(FluidTags.WATER) || z < y) {
 	            aa = this.getHeartRows(x) - 1;
 	            t -= aa * 10;
-	            ab = MathHelper.ceil((double)((double)(z - 2) * 10.0 / (double)y));
-	            ac = MathHelper.ceil((double)((double)z * 10.0 / (double)y)) - ab;
+	            ab = MathHelper.ceil((double)(z - 2) * 10.0 / (double)y);
+	            ac = MathHelper.ceil((double)z * 10.0 / (double)y) - ab;
 	            for (int ad = 0; ad < ab + ac; ++ad) {
 	                if (ad < ab) {
 	                    dc.drawTexture(ICONS, n - ad * 8 - 9, t, 16, 18, 9, 9);
@@ -235,8 +235,8 @@ public class RenderOverlayMixin {
     private void renderHealthBar(DrawContext context, Random random, PlayerEntity player, int x, int y, int lines, int regeneratingHeartIndex, float maxHealth, int lastHealth, int health, int absorption, boolean blinking) {
         HeartTypeNew heartType = HeartTypeNew.fromPlayerState(player);
         int i = 9 * (player.getWorld().getLevelProperties().isHardcore() ? 5 : 0);
-        int j = MathHelper.ceil((double)((double)maxHealth / 2.0));
-        int k = MathHelper.ceil((double)((double)absorption / 2.0));
+        int j = MathHelper.ceil((double)maxHealth / 2.0);
+        int k = MathHelper.ceil((double)absorption / 2.0);
         int l = j * 2;
         for (int m = j + k - 1; m >= 0; --m) {
             boolean bl3;
